Tidied up API app
