+++ data/mysql/binlog.index
@@ -4,3 +4,4 @@
 ./binlog.000014
 ./binlog.000015
 ./binlog.000016
+./binlog.000017
