+++ src/api/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Source, Graph, GraphDataset
+from api.models import Source, Graph, GraphDataset
 
 admin.site.register(Source)
 admin.site.register(Graph)
